chore: Moved log-level.enum.mts to top domain definition folder.
