--- packages/architectura/src/Service/Logger/index.mts
@@ -1,1 +1,0 @@
-export * from "../stack-trace-parser/stack-trace-parser.service.mjs";
