+++ packages/architectura/src/definition/enum/index.mts
@@ -1,1 +1,3 @@
+export * from './log-level.enum.mjs';
 export * from './stack-trace-parser.enum.mjs';
+
