+++ packages/architectura/src/definition/enum/log-level.enum.mts
